Move error formatting to fmt.rs
