+++ src/error.rs
@@ -652,16 +652,16 @@ impl<E> ErrorImpl<E> {
 }
 
 impl ErrorImpl<()> {
-    fn error(&self) -> &(dyn StdError + Send + Sync + 'static) {
+    pub(crate) fn error(&self) -> &(dyn StdError + Send + Sync + 'static) {
         unsafe { &*(self.vtable.object_ref)(self) }
     }
 
-    fn error_mut(&mut self) -> &mut (dyn StdError + Send + Sync + 'static) {
+    pub(crate) fn error_mut(&mut self) -> &mut (dyn StdError + Send + Sync + 'static) {
         unsafe { &mut *(self.vtable.object_mut)(self) }
     }
 
     #[cfg(backtrace)]
-    fn backtrace(&self) -> &Backtrace {
+    pub(crate) fn backtrace(&self) -> &Backtrace {
         // This unwrap can only panic if the underlying error's backtrace method
         // is nondeterministic, which would only happen in maliciously
         // constructed code.
@@ -671,71 +671,10 @@ impl ErrorImpl<()> {
             .expect("backtrace capture failed")
     }
 
-    fn chain(&self) -> Chain {
+    pub(crate) fn chain(&self) -> Chain {
         Chain {
             next: Some(self.error()),
         }
-    }
-
-    fn print_chain(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        write!(f, "{}", self.error())?;
-
-        let mut chain = self.chain().skip(1).enumerate().peekable();
-        if let Some((n, error)) = chain.next() {
-            write!(f, "\n\nCaused by:\n    ")?;
-            if chain.peek().is_some() {
-                write!(f, "{}: ", n)?;
-            }
-            write!(f, "{}", error)?;
-            for (n, error) in chain {
-                write!(f, "\n    {}: {}", n, error)?;
-            }
-        }
-
-        Ok(())
-    }
-
-    fn display(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        if f.alternate() {
-            self.print_chain(f)
-        } else {
-            write!(f, "{}", self.error())
-        }
-    }
-
-    fn debug(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        if f.alternate() {
-            return Debug::fmt(self.error(), f);
-        }
-
-        self.print_chain(f)?;
-        writeln!(f)?;
-
-        #[cfg(backtrace)]
-        {
-            use std::backtrace::BacktraceStatus;
-
-            let backtrace = self.backtrace();
-            match backtrace.status() {
-                BacktraceStatus::Captured => {
-                    let mut backtrace = backtrace.to_string();
-                    if backtrace.starts_with("stack backtrace:") {
-                        // Capitalize to match "Caused by:"
-                        backtrace.replace_range(0..1, "S");
-                    }
-                    write!(f, "\n{}", backtrace)?;
-                }
-                BacktraceStatus::Disabled => {
-                    writeln!(
-                        f,
-                        "\nStack backtrace:\n    Run with RUST_LIB_BACKTRACE=1 env variable to display a backtrace"
-                    )?;
-                }
-                _ => {}
-            }
-        }
-
-        Ok(())
     }
 }
